공통 모듈 모듈화하여 중복 제거
