+++ src/games/game/Game.module.css
@@ -8,7 +8,7 @@
 
 .imageWrapper {
   position: relative;
-  padding-top: 56.25%; /* 720 / 1280 = 0.5625 */
+  padding-top: 56.25%;
 }
 
 .imageWrapper > div {
