+++ styles/game/game.module.css
@@ -23,7 +23,7 @@
 
 .playerWrapper {
   position: relative;
-  padding-top: 56.25%; /* 720 / 1280 = 0.5625 */
+  padding-top: 56.25%;
   margin-bottom: 1em;
 }
